--- src/main/java/com/example/backendmentoring/article/ArticleRepository.java
@@ -1,7 +1,0 @@
-package com.example.backendmentoring.article;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ArticleRepository extends JpaRepository<Article, Long> {
-
-}
