+++ src/main/java/com/example/backendmentoring/article/application/CreateArticleUseCase.java
@@ -1,0 +1,9 @@
+package com.example.backendmentoring.article.application;
+
+import com.example.backendmentoring.article.dto.ArticleRequestDto;
+import com.example.backendmentoring.article.dto.ArticleResponseDto;
+
+public interface CreateArticleUseCase {
+
+    ArticleResponseDto createArticle(ArticleRequestDto requestDto, String userEmail);
+}
