+++ src/main/java/com/example/backendmentoring/article/domain/ArticleRepository.java
@@ -1,0 +1,8 @@
+package com.example.backendmentoring.article.domain;
+
+import com.example.backendmentoring.article.Article;
+
+public interface ArticleRepository {
+
+    Article save(Article article);
+}
